ai_services/speech_vision/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
transformers==4.44.2
torch==2.4.1
Pillow==10.4.0
requests==2.32.3
